common TableArg
